+++ actix-web-httpauth/src/headers/authorization/scheme/mod.rs
@@ -1,6 +1,6 @@
 use std::fmt::{Debug, Display};
 
-use actix_web::http::header::{HeaderValue, IntoHeaderValue};
+use actix_web::http::header::{HeaderValue, TryIntoHeaderValue};
 
 pub mod basic;
 pub mod bearer;
@@ -9,7 +9,7 @@ use crate::headers::authorization::errors::ParseError;
 
 /// Authentication scheme for [`Authorization`](./struct.Authorization.html)
 /// header.
-pub trait Scheme: IntoHeaderValue + Debug + Display + Clone + Send + Sync {
+pub trait Scheme: TryIntoHeaderValue + Debug + Display + Clone + Send + Sync {
     /// Try to parse the authentication scheme from the `Authorization` header.
     fn parse(header: &HeaderValue) -> Result<Self, ParseError>;
 }