+++ actix-web-httpauth/src/headers/www_authenticate/challenge/basic.rs
@@ -5,7 +5,7 @@ use std::default::Default;
 use std::fmt;
 use std::str;
 
-use actix_web::http::header::{HeaderValue, IntoHeaderValue, InvalidHeaderValue};
+use actix_web::http::header::{HeaderValue, InvalidHeaderValue, TryIntoHeaderValue};
 use actix_web::web::{BufMut, Bytes, BytesMut};
 
 use super::Challenge;
@@ -103,18 +103,17 @@ impl fmt::Display for Basic {
     }
 }
 
-impl IntoHeaderValue for Basic {
+impl TryIntoHeaderValue for Basic {
     type Error = InvalidHeaderValue;
 
-    fn try_into_value(self) -> Result<HeaderValue, <Self as IntoHeaderValue>::Error> {
+    fn try_into_value(self) -> Result<HeaderValue, Self::Error> {
         HeaderValue::from_maybe_shared(self.to_bytes())
     }
 }
 
 #[cfg(test)]
 mod tests {
-    use super::Basic;
-    use actix_web::http::header::IntoHeaderValue;
+    use super::*;
 
     #[test]
     fn test_plain_into_header_value() {