+++ actix-session/CHANGES.md
@@ -6,6 +6,7 @@
 * `SessionMiddleware`, a middleware to provide support for saving/updating/deleting session state against a pluggable storage backend (see `SessionStore` trait) [#212]
 * `CookieSessionStore`, a cookie-based backend to store session state [#212]
 * `RedisActorSessionStore`, a Redis-based backend to store session state powered by `actix-redis` [#212]
+* `RedisSessionStore`, a Redis-based backend to store session state powered by `redis-rs` [#212]
 
 ### Changed
 * Rename `UserSession` to `SessionExt` [#212]