+++ actix-session/CHANGES.md
@@ -3,6 +3,12 @@
 ## Unreleased - 2021-xx-xx
 
 
+## 0.5.0-beta.6 - 2021-12-18
+* Update `actix-web` dependency to `4.0.0.beta-15`. [#216]
+
+[#216]: https://github.com/actix/actix-extras/pull/216
+
+
 ## 0.5.0-beta.5 - 2021-12-12
 * Update `actix-web` dependency to `4.0.0.beta-14`. [#209]
 * Remove `UserSession` implementation for `RequestHead`. [#209]